Shorten Array's size from 32 to 24
The `Array::cap` is the same as `Array::arr.len()` so we don't need it.
By removing it, we reduce Array's size from 32 to 24. Correspondingly,
`Representation` is shortened from 40 to 32.

This is important for our workloads where vast majority of things have
low cardinality and are often represented by `Array`. Although `Array`
is just an itermediate representation, it still helps in reducing temp
memory usage.